refactor(farmer): split FarmerCopilot (615→166) em hook + componentes
Extrai a lógica (sessão, transcrição ElevenLabs scribe, prepareSessionContext,
handlers de voz/texto/análise/cópia, derivações) para o hook useFarmerCopilot e
quebra o JSX em SessionStartCard, DirectionIndicator, SuggestionCard,
ActivePlanCard, ManualTextInput, TranscriptCard e AnalysisHistoryCard. Constantes
em config.ts, InputMode em types.ts. Comportamento 1:1; +15 testes.
