+++ src/components/farmer/copilot/types.ts
@@ -1,0 +1,4 @@
+// Tipos do copiloto de vendas.
+// Extraídos verbatim de src/pages/FarmerCopilot.tsx (god-component split).
+
+export type InputMode = 'voice' | 'text';
